musl: improved TCP socket, works in happy day scenarios

+++ api/posix/udp_fd.hpp
@@ -46,7 +46,7 @@ public:
 
   /** SOCKET */
   long    bind(const struct sockaddr *, socklen_t) override;
-  int     connect(const struct sockaddr *, socklen_t) override;
+  long    connect(const struct sockaddr *, socklen_t) override;
 
   ssize_t sendto(const void *, size_t, int, const struct sockaddr *, socklen_t) override;
